--- src/ui/styled/CssBaseline/constant.ts
@@ -1,1 +1,0 @@
-export const SPACING_BASE_PX = 8;
