+++ src/ui/styled/constant.ts
@@ -1,1 +1,2 @@
 export const SrOnlyClassName = 'sr-only';
+export const SPACING_BASE_PX = 8;
