+++ src/ui/styled/variables/space.ts
@@ -1,5 +1,5 @@
 import { pxToRem } from '@/ui/lib/fonts';
-import { SPACING_BASE_PX } from '@/ui/styled/CssBaseline/constant';
+import { SPACING_BASE_PX } from '@/ui/styled/constant';
 
 export const Spaces = ['½', 1, 2, 3, 4, 5, 6];
 export type SpaceGap = (typeof Spaces)[number];
